REFACTOR commands to src/commands directory
