+++ src/Commands/Semver/BumpPatch.php
@@ -1,26 +1,26 @@
 <?php
 
-namespace Sfneal\Cruise\Semver;
+namespace Sfneal\Cruise\Commands\Semver;
 
 use Exception;
 use Illuminate\Console\Command;
 use Illuminate\Support\Facades\Artisan;
 
-class BumpMinor extends Command
+class BumpPatch extends Command
 {
     /**
      * The name and signature of the console command.
      *
      * @var string
      */
-    protected $signature = 'bump:minor';
+    protected $signature = 'bump:patch';
 
     /**
      * The console command description.
      *
      * @var string
      */
-    protected $description = 'Bump the application to the next minor version (e.g. v1.1 to v1.2)';
+    protected $description = 'Bump the application to the next patch version (e.g. v1.3.4 to v1.3.5)';
 
     /**
      * Execute the console command.
@@ -30,6 +30,6 @@ class BumpMinor extends Command
      */
     public function handle(): int
     {
-        return Artisan::call('bump minor');
+        return Artisan::call('bump patch');
     }
 }
